Move deploy status out of gulp
